Split SRC_URI list into inc file
Usually generated bitbake recipes need to be slightly altered before
they can be used to build the rust project. This might be tweaking some
URLs, adding non-rust dependencies, and a variety of other things.
Typically these changes will be made to the generated recipe and then
checked in to a yocto layer.

As the rust project changes it will usually pull in additional crates
over time. Whenever this happens the recipe will be re-generated and
these changes will need to be merged back in. This is a lot of work to
do when the only change is the addition of one or two crates to the
SRC_URI list.

Let's make this process a little easier by splitting the SRC_URI list
out into a separate inc file and including this in the main recipe. Now
the recipe can be kept up-to-date by just updating the separate inc
file--no more recipe merging.